Develop To QA  (#602)
* feat: support multiple did methods (#589)

* feat: support multiple did methods


* refactor: removed the commented code


* fix: removed commented code


* fix: removed commented code agent-service


* fix: removed commented code agent-service


---------


* fix: refactored tables


* fix: reafctored json data


* feat:add reuse connection


* fix: refactored tables (#592)

* fix: refactored tables


* fix: reafctored json data


---------


* fix:changed validation message


* worked on the POST API of refresh token


* fix: an incorrect message appears when delete the ecosystem invitation which is already deleted


* fix: schema id validations


* feat: dedicated agent spin up


* fix: schema validation


* fix: schema validations


---------

+++ apps/ecosystem/src/ecosystem.repository.ts
@@ -868,6 +868,18 @@ export class EcosystemRepository {
   // eslint-disable-next-line camelcase
   async deleteInvitations(invitationId: string): Promise<ecosystem_invitations> {
     try {
+
+      const findInvitation = await this.prisma.ecosystem_invitations.findUnique({
+        where: {
+          id: invitationId,
+          status: EcosystemInvitationStatus.PENDING
+        }
+      });
+      
+      if (!findInvitation) {
+        throw new NotFoundException('Ecosystem Invitation not found');
+      }
+
       const deletedInvitation = await this.prisma.ecosystem_invitations.delete({
         where: {
           id: invitationId,
